fixed trivia code rendering on other pages besides home, trivia is now its own component, code is a lot more readable
