+++ src/components/Homepage/Homepage.jsx
@@ -1,17 +1,20 @@
 import React from "react";
-import styles from "./Homepage.module.css"; 
+import styles from "./Homepage.module.css";
+import Trivia from '../Trivia/Trivia'; 
 
 function Homepage() {
   return (
-    <div className={styles.homepage}> {/* Use the styles from the CSS module */}
+    <div className={styles.homepage}>
       <h2 className={styles["player-name"]}>Player's Name</h2>
       <div className={styles["category-options"]}>
         {/* Include the category dropdown/select and number of questions */}
         {/* Include the "Play" button */}
         <button className={styles["play-button"]}>Play</button>
       </div>
+      <Trivia /> {/* Render the Trivia component */}
     </div>
   );
 }
 
 export default Homepage;
+
