+++ src/components/Trivia/GameResult.jsx
